restructure as poetry project
