+++ pyproject.toml
@@ -1,0 +1,22 @@
+[tool.poetry]
+name = "uptime-service-validation"
+version = "0.1.0"
+description = "Mina Delegation Program Coordinator and Validator Component"
+authors = ["Mina Foundation"]
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = ">=3.10,<3.13"
+kubernetes = "^28.1.0"
+python-dotenv = "^1.0.0"
+slackclient = "^2.9.4"
+boto3 = "^1.28.83"
+pandas = "^2.1.2"
+networkx = "^3.2.1"
+matplotlib = "^3.8.1"
+psycopg2 = "^2.9.9"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
